bootstraprag/templates/evaluations/phoenix_evals/requirements.txt
arize-phoenix==5.2.2
python-dotenv==1.0.1
pandas==2.2.3